+++ staticmodel/django/rest_framework/serializers.py
@@ -3,30 +3,32 @@
 Django Rest Framework serializer fields
 ***************************************
 
-**Static Model** provides custom DRF (2.x) serializer fields in the
+**Static Model** provides custom serializer fields in the
 ``staticmodel.django.rest_framework.serializers`` module:
 
  * ``StaticModelCharField`` (sub-class of ``rest_framework.serializers.CharField``)
  * ``StaticModelIntegerField`` (sub-class of ``rest_framework.serializers.IntegerField``)
 
-Currently, only a single field value is extracted from the static model
-instance when serializing. Support for serializing the entire instance
-as a nested dict will be added in the future.
-
-When deserializing, the single field value is looked up in the static model
-and the corresponding member is returned.
-
 All fields take the following keyword arguments in addition to the
 arguments taken by their respective parent classes:
 
  * ``static_model``: The static model class associated with this field.
- * ``lookup_field_name``: The static model field name that will be
-   returned during serialization and that will be used to lookup the
-   static model member when deserializing. Defaults to the first field
-   name in ``static_model._field_names``.
+ * ``lookup_field_name``: The static model field name that will be used
+   to lookup the static model member when deserializing, and the field
+   name to retrieve the value from when serializing (unless
+   ``static_model_expand=True``. See below.). Defaults to the first field
+    name in ``static_model._field_names``.
+ * ``static_model_expand``: When set to ``True``, return the entire
+    static model member as a mapping. Defaults to ``False``.
+
+Regardless of the value of ``static_model_expand``, if the value passed
+during deserialization is a mapping, it will be used to retrieve the
+lookup value using ``lookup_field_name``.
+
 """
-from django.core.exceptions import ValidationError
+from collections.abc import Mapping
 from rest_framework import serializers
+from rest_framework.exceptions import ValidationError
 from staticmodel import StaticModel
 
 
@@ -48,24 +50,37 @@ class StaticModelFieldMixin(object):
             lookup_field_name = self._static_model._field_names[0]
         self._lookup_field_name = lookup_field_name
 
+        self._expand = kwargs.pop('static_model_expand', False)
+
         super().__init__(*args, **kwargs)
 
-    def to_native(self, value):
+    def to_representation(self, value):
         if value is None:
             return value
         elif isinstance(value, self._static_model):
-            return getattr(value, self._lookup_field_name)
+            if self._expand:
+                return dict(value._as_dict)
+            else:
+                return getattr(value, self._lookup_field_name)
         else:
             raise ValueError("Invalid value for 'value' parameter")
 
-    def from_native(self, value):
-        if value is None:
-            return value
+    def to_internal_value(self, data):
+        if data is None:
+            return data
         else:
+            if isinstance(data, Mapping):
+                try:
+                    lookup_value = data[self._lookup_field_name]
+                except KeyError:
+                    raise ValidationError("Representation missing field '{}'".format(
+                        self._lookup_field_name))
+            else:
+                lookup_value = data
             try:
-                return self._static_model.members.get(**{self._lookup_field_name: value})
+                return self._static_model.members.get(**{self._lookup_field_name: lookup_value})
             except self._static_model.DoesNotExist as e:
-                raise ValidationError("Value {!r} is invalid".format(value))
+                raise ValidationError("Value {!r} is invalid".format(data))
 
 
 class StaticModelCharField(StaticModelFieldMixin, serializers.CharField):